@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'active_support/concern'
 
 module ClosureTree
@@ -28,11 +30,12 @@ module ClosureTree
 
     def _ct_validate
       if !(defined? @_ct_skip_cycle_detection) &&
-        !new_record? && # don't validate for cycles if we're a new record
-        changes[_ct.parent_column_name] && # don't validate for cycles if we didn't change our parent
-        parent.present? && # don't validate if we're root
-        parent.self_and_ancestors.include?(self) # < this is expensive :\
-        errors.add(_ct.parent_column_sym, I18n.t('closure_tree.loop_error', default: 'You cannot add an ancestor as a descendant'))
+         !new_record? && # don't validate for cycles if we're a new record
+         changes[_ct.parent_column_name] && # don't validate for cycles if we didn't change our parent
+         parent.present? && # don't validate if we're root
+         parent.self_and_ancestors.include?(self) # < this is expensive :\
+        errors.add(_ct.parent_column_sym,
+                   I18n.t('closure_tree.loop_error', default: 'You cannot add an ancestor as a descendant'))
       end
     end
 
@@ -51,11 +54,8 @@ module ClosureTree
         return
       end
 
-      if public_send(:saved_changes)[_ct.parent_column_name] || @was_new_record
-        rebuild!
-      end
-
-      if public_send(:saved_changes)[_ct.parent_column_name] && !@was_new_record
+      rebuild! if saved_changes[_ct.parent_column_name] || @was_new_record
+      if saved_changes[_ct.parent_column_name] && !@was_new_record
         # Resetting the ancestral collections addresses
         # https://github.com/mceachen/closure_tree/issues/68
         ancestor_hierarchies.reload
@@ -69,9 +69,7 @@ module ClosureTree
     def _ct_before_destroy
       _ct.with_advisory_lock do
         delete_hierarchy_references
-        if _ct.options[:dependent] == :nullify
-          self.class.find(self.id).children.find_each { |c| c.rebuild! }
-        end
+        self.class.find(id).children.find_each(&:rebuild!) if _ct.options[:dependent] == :nullify
       end
       true # don't prevent destruction
     end
@@ -94,7 +92,7 @@ module ClosureTree
     def rebuild!(called_by_rebuild = false)
       _ct.with_advisory_lock do
         delete_hierarchy_references unless was_new_record?
-        hierarchy_class.create!(:ancestor => self, :descendant => self, :generations => 0)
+        hierarchy_class.create!(ancestor: self, descendant: self, generations: 0)
         unless root?
           _ct.connection.execute <<-SQL.squish
             INSERT INTO #{_ct.quoted_hierarchy_table_name}
@@ -108,7 +106,7 @@ module ClosureTree
         if _ct.order_is_numeric? && !@_ct_skip_sort_order_maintenance
           _ct_reorder_prior_siblings_if_parent_changed
           # Prevent double-reordering of siblings:
-          _ct_reorder_siblings if !called_by_rebuild
+          _ct_reorder_siblings unless called_by_rebuild
         end
 
         children.find_each { |c| c.rebuild!(true) }
@@ -125,16 +123,10 @@ module ClosureTree
         # It shouldn't affect performance of postgresql.
         # See http://dev.mysql.com/doc/refman/5.0/en/subquery-errors.html
         # Also: PostgreSQL doesn't support INNER JOIN on DELETE, so we can't use that.
-        _ct.connection.execute <<-SQL.squish
-          DELETE FROM #{_ct.quoted_hierarchy_table_name}
-          WHERE descendant_id IN (
-            SELECT DISTINCT descendant_id
-            FROM (SELECT descendant_id
-              FROM #{_ct.quoted_hierarchy_table_name}
-              WHERE ancestor_id = #{_ct.quote(id)}
-                 OR descendant_id = #{_ct.quote(id)}
-            ) #{ _ct.t_alias_keyword } x )
-        SQL
+
+        hierarchy_table = hierarchy_class.arel_table
+        delete_query = _ct.build_hierarchy_delete_query(hierarchy_table, id)
+        _ct.connection.execute(delete_query.to_sql)
       end
     end
 
@@ -142,7 +134,7 @@ module ClosureTree
       (defined? @was_new_record) && @was_new_record
     end
 
-    module ClassMethods
+    class_methods do
       # Rebuilds the hierarchy table based on the parent_id column in the database.
       # Note that the hierarchy table will be truncated.
       def rebuild!
@@ -156,8 +148,8 @@ module ClosureTree
       def cleanup!
         hierarchy_table = hierarchy_class.arel_table
 
-        [:descendant_id, :ancestor_id].each do |foreign_key|
-          alias_name = foreign_key.to_s.split('_').first + "s"
+        %i[descendant_id ancestor_id].each do |foreign_key|
+          alias_name = "#{foreign_key.to_s.split('_').first}s"
           alias_table = Arel::Table.new(table_name).alias(alias_name)
           arel_join = hierarchy_table.join(alias_table, Arel::Nodes::OuterJoin)
                                      .on(alias_table[primary_key].eq(hierarchy_table[foreign_key]))
