@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module ClosureTree
   module HashTree
     extend ActiveSupport::Concern
@@ -6,8 +8,7 @@ module ClosureTree
       _ct.hash_tree(self_and_descendants, options[:limit_depth])
     end
 
-    module ClassMethods
-
+    class_methods do
       # There is no default depth limit. This might be crazy-big, depending
       # on your tree shape. Hash huge trees at your own peril!
       def hash_tree(options = {})
