@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module ClosureTree
-  VERSION = Gem::Version.new('8.0.0')
+  VERSION = Gem::Version.new('9.1.1')
 end
