@@ -1,8 +1,9 @@
 #!/usr/bin/env bundle exec ruby -I lib:spec
+# frozen_string_literal: true
 
 # Simple benchmark utility to create a closure tree based on the topology of the current filesystem
 
-#ENV['NONUKES'] = '1'
+# ENV['NONUKES'] = '1'
 require 'spec_helper'
 require 'findler'
 require 'pathname'
@@ -22,8 +23,8 @@ f = Findler.new '/'
 iter = f.iterator
 Tag.with_advisory_lock('closure_tree') do
   while (nxt = iter.next_file) && ((cnt += 1) < 1000)
-    t = Tag.find_or_create_by_path(nxt.path_array)
-    puts "created #{nxt.to_s}"
+    Tag.find_or_create_by_path(nxt.path_array)
+    puts "created #{nxt}"
   end
 end
 
@@ -31,8 +32,7 @@ puts "Tag.all.size: #{Tag.all.size}"
 puts "TagHierarchy.all.size: #{TagHierarchy.all.size}"
 
 puts 'Tag.roots performance:'
-puts Benchmark.measure { Tag.roots.size }
+puts(Benchmark.measure { Tag.roots.size })
 
 puts 'Tag.leaves performance:'
-puts Benchmark.measure { Tag.leaves.size }
-
+puts(Benchmark.measure { Tag.leaves.size })
