@@ -1,7 +1,9 @@
 # frozen_string_literal: true
 
 require 'test_helper'
+require 'support/tag_examples'
 
-describe Tag do
+class TagTest < ActiveSupport::TestCase
+  TAG_CLASS = Tag
   include TagExamples
 end
