@@ -1,7 +1,9 @@
 # frozen_string_literal: true
 
 require 'test_helper'
+require 'support/tag_examples'
 
-describe UUIDTag do
+describe UuidTag do
+  TAG_CLASS = UuidTag
   include TagExamples
 end
