@@ -3,21 +3,22 @@
 # https://stackoverflow.com/a/43810063/1683557
 
 module QueryCounter
-  def sql_queries(&block)
+  def sql_queries(&)
     queries = []
     counter = lambda { |*, payload|
       queries << payload.fetch(:sql) unless %w[CACHE SCHEMA].include?(payload.fetch(:name))
     }
 
-    ActiveSupport::Notifications.subscribed(counter, "sql.active_record", &block)
+    ActiveSupport::Notifications.subscribed(counter, 'sql.active_record', &)
 
     queries
   end
 
-  def assert_database_queries_count(expected, &block)
-    queries = sql_queries(&block)
-    queries.count.must_equal(
+  def assert_database_queries_count(expected, &)
+    queries = sql_queries(&)
+    assert_equal(
       expected,
+      queries.count,
       "Expected #{expected} queries, but found #{queries.count}:\n#{queries.join("\n")}"
     )
   end
